requirements.txt
asgiref==3.2.7
attrs==19.3.0
Django==3.0.6
djangorestframework==3.11.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
wcwidth==0.1.9
